[FEATURE] Support for composer commands
Changes the commands to composer commands so each can now be executed by composer too. Usage as symfony base application remains unchainged.
