+++ README.md
@@ -2,8 +2,8 @@
 
 ## Install
 
-```
-$ composer require bk2k/extension-helper --dev
+```bash
+composer require bk2k/extension-helper --dev
 ```
 
 ## Commands
@@ -13,9 +13,9 @@ $ composer require bk2k/extension-helper --dev
 This is a wrapper command that calls different commands `version:set`,
 `changelog:create`, `release:publish` and `archive:create` in sequence.
 
-```
-$ php bin/extension-helper release:create <next version number>
-$ php bin/extension-helper release:create 1.0.0
+```bash
+composer release:create <next version number>
+composer release:create 1.0.0
 ```
 
 ### `release:publish`
@@ -23,18 +23,18 @@ $ php bin/extension-helper release:create 1.0.0
 This command will commit all uncommitted files in the working directory,
 and adds the version number as tag.
 
-```
-$ php bin/extension-helper release:publish <next version number>
-$ php bin/extension-helper release:publish 1.0.0
+```bash
+composer release:publish <next version number>
+composer release:publish 1.0.0
 ```
 
 ### `changelog:create`
 
 This command generates a changelog from your git log.
 
-```
-$ php bin/extension-helper changelog:create <next version number>
-$ php bin/extension-helper changelog:create 1.0.0
+```bash
+composer changelog:create <next version number>
+composer changelog:create 1.0.0
 ```
 
 ### `version:set`
@@ -46,18 +46,16 @@ For now it will update files:
 - `Documentation\Settings.yaml`
 - `ext_emconf.php`
 
-```
-$ php bin/extension-helper version:set <next version number> --dev
+```bash
+composer version:set <next version number> --dev
 ```
 
-```
+```bash
 # Set version to 1.0.0
-$ php bin/extension-helper version:set 1.0.0
-```
+composer version:set 1.0.0
 
-```
 # Set version to 1.0.0-dev
-$ php bin/extension-helper version:set 1.0.0 --dev
+composer version:set 1.0.0 --dev
 ```
 
 ### `archive:create`
@@ -68,10 +66,19 @@ With the optional parameter you can set a specific version number the archive
 should be created for. The script will fail if you provide a version number
 and the tag does not exist.
 
-```
+```bash
 # packagename_<branch>-<revision>.zip
-$ php bin/extension-helper archive:create
+composer archive:create
 
 # packagename_1.0.0.zip
-$ php bin/extension-helper archive:create 1.0.0
+composer archive:create 1.0.0
+```
+
+### Legacy usage
+
+Old style usage is still working as before by calling `php bin/extension-helper`
+instead of `composer`.
+
+```bash
+php bin/extension-helper release:create 1.0.0
 ```
