+++ src/Composer/Plugin.php
@@ -1,0 +1,34 @@
+<?php
+declare(strict_types=1);
+
+/*
+ * This file is part of the bk2k/extension-helper.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE file that was distributed with this source code.
+ */
+
+namespace BK2K\ExtensionHelper\Composer;
+
+use BK2K\ExtensionHelper\Composer\Command;
+use Composer\Composer;
+use Composer\IO\IOInterface;
+use Composer\Plugin\PluginInterface;
+use Composer\Plugin\Capable;
+
+/**
+ * Plugin
+ */
+class Plugin implements PluginInterface, Capable
+{
+    public function activate(Composer $composer, IOInterface $io)
+    {
+    }
+
+    public function getCapabilities()
+    {
+        return [
+            'Composer\Plugin\Capability\CommandProvider' => 'BK2K\ExtensionHelper\Composer\Commands',
+        ];
+    }
+}
